requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-debug-toolbar==1.3.0
django-extensions==1.5.0
django-secure==1.0.1
gunicorn==19.2.1
psycopg2==2.6
Werkzeug==0.10.1

django-zurb-foundation==5.5.0
django-fontawesome==0.2.2

# requests must be installed before github3 for this to install properly
requests==2.5.3
github3.py==1.0.0a1
celery==3.1.18
django-celery==3.1.16
#highlander-one==0.1.1
amqp==1.4.6
anyjson==0.3.3

ipython
